+++ src/components/Note.css
